- Moved more scan classes into the scanning package
