+++ src/burp/model/scanning/NormalisedParamScan.java
@@ -1,16 +1,17 @@
-package burp;
+package burp.model.scanning;
 
-import burp.model.scanning.ParamScan;
+import burp.IHttpRequestResponse;
+import burp.IHttpService;
+import burp.IScanIssue;
+import burp.IScannerInsertionPoint;
 import burp.albinowaxUtils.Resp;
-import burp.model.scanning.Scan;
-import burp.model.scanning.BulkScanLauncher;
 import burp.model.utilities.Utilities;
 
 import java.util.List;
 
 public class NormalisedParamScan extends ParamScan {
 
-    NormalisedParamScan(String name, Utilities utilities, BulkScanLauncher launcher) {
+    public NormalisedParamScan(String name, Utilities utilities, BulkScanLauncher launcher) {
         super(name, utilities, launcher);
     }
